requirements.txt
fastapi==0.101.0
uvicorn==0.21.1
google-ai-generativelanguage==0.6.9
google-api-core==2.20.0
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.1
google-pasta==0.2.0
googleapis-common-protos==1.65.0
python-dotenv==1.0.1
PyPDF2==3.0.1
pandas==2.0.3
bs4==0.0.2
numpy==1.26.4
celery==5.2.7